move test to separate file
